feat: Set client id as long

+++ app/src/main/java/com/example/hueverianieto/ui/views/clients/fragments/newclient/NewClientFragment.kt
@@ -183,7 +183,7 @@ class NewClientFragment : BaseFragment() {
                         direction,
                         email,
                         hasAccount,
-                        nextIdStr,
+                        nextIdStr.toLong(),
                         listOf(
                             mapOf(namePhone1 to phone1.toLong()),
                             mapOf(namePhone2 to phone2.toLong())
